Configure HornetQ to internally retry failed connections
This fixes an issue where a time jump would cause the HornetQ client to decide
that the connection failed and, with the default configuration, just plain
give up.

Presumably it expected application level retry to be implemented, but that
wasn't present in TorqueBox, so nothing happened on the queue again.

This change turns on the internal retry policy in the client; perhaps less
pleasant to the upper layers attempts to recover, but more effective when they
simply ignore the failure.

Fixes https://tickets.puppetlabs.com/browse/RAZOR-644