Revert the tarball to continue hard coding version
Unfortunately, in this automation world, we parse the version of the
package we're building from the tag of the project. I have yet to figure
out an easy way to simply use the original tag, i.e. `3.1.1`, rather
than the derived version, i.e., `3.1.1.2`, which is created during the
package build. Since we need the version in order to correctly refer to
the tarball, we need this information somewhere. I put it here.